pyproject.toml
[project]
name = "imgprockit"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Samurai-Human-Go_Jetson", email = "gotechsamurai@samurai-human-go.com" }
]
dependencies = [
    "pillow>=11.3.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/imgprockit"]
